examples/autoclick/requirements.txt
aiohttp
beautifulsoup4
lxml
python-dotenv
